moving get_default_jupyter_workdir to the utils.py
